double size of premises

+++ src/main/java/sda/covidrestrictions/tracker/Catering.java
@@ -39,8 +39,8 @@ public class Catering {
     public static void getNumberOfPeopleToOrderTakeAway() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Please, enter the size of your premises used for customer service in square meters: ");
-        int sizeOfPremises = scanner.nextInt();
-        if (sizeOfPremises > 49) numberOfPeople = sizeOfPremises / allowedSqMetersPerPerson;
+        double sizeOfPremises = scanner.nextDouble();
+        if (sizeOfPremises > 49) numberOfPeople = (int) (sizeOfPremises / allowedSqMetersPerPerson);
         System.out.println("Allowed number of customers in your shop is : " + numberOfPeople);
 
 }
